+++ src/ralph/__init__.py
@@ -1,1 +1,1 @@
-VERSION = ('2', '0', '0-rc8')
+VERSION = ('2', '0', '0-stable')