+++ src/ralph/discovery/tests/plugins/test_xen.py
@@ -171,10 +171,6 @@ uuid ( RO)             : 9393c492-4311-4dde-9413-75150afdcd97
 \_ 2:0:2:8 sdi 8:128 [active][undef]
 \_ round-robin 0 [prio=0][enabled]
 \_ 1:0:2:8 sdh 8:112 [active][undef]
-3600508b1001c7e8cccd13c7e7a61730f dm-0 HP,LOGICAL VOLUME
-[size=137G][features=0][hwhandler=0][rw]
-\_ round-robin 0 [prio=0][active]
-\_ 0:0:0:1 sda 8:0   [active][undef]
 350002ac01add042f dm-4 3PARdata,VV
 [size=1000G][features=1 queue_if_no_path][hwhandler=0][rw]
 \_ round-robin 0 [prio=0][active]
@@ -215,7 +211,7 @@ uuid ( RO)             : 9393c492-4311-4dde-9413-75150afdcd97
   060d5424-b9f1-ad56-629b-10c0c4e2cfa5     XSLocalEXT-060d5424-b9f1-ad56-629b-10c0c4e2cfa5    -wi-ao 128.68M
 """),
         ])
-        shares = hardware.get_disk_shares(ssh)
+        shares = hardware.get_disk_shares(ssh, include_logical_volumes=True)
         self.maxDiff = None
         self.assertEqual(shares, {
             'MGT': ('600144F008BF8A000000500CFFF20003', 4),