+++ src/ralph/settings.py
@@ -18,6 +18,7 @@ execfile(namespace_package_support)
 #
 # common stuff for each install
 #
+from datetime import timedelta
 
 SITE_ID = 1
 USE_I18N = True
@@ -125,13 +126,14 @@ LOGGING = {
     },
     'formatters': {
         'verbose': {
-            'datefmt': '%H:%M:%S',
+            'datefmt': '%d.%m.%Y %H:%M:%S',
             'format': (
-                '%(asctime)08s,%(msecs)03d %(levelname)-7s [%(processName)s'
+                '[%(asctime)08s,%(msecs)03d] %(levelname)-7s [%(processName)s'
                 ' %(process)d] %(module)s - %(message)s'),
         },
         'simple': {
-            'format': '%(levelname)s %(message)s',
+            'datefmt': '%H:%M:%S',
+            'format': '[%(asctime)08s] %(levelname)-7s %(message)s',
         },
     },
     'loggers': {
@@ -152,6 +154,7 @@ LOGGING = {
         },
     },
 }
+POSSIBLE_EVENTS_TIMEDELTA = timedelta(days=30)
 FORCE_SCRIPT_NAME = ''
 # testing settings
 import os
@@ -363,10 +366,7 @@ DEAD_PING_COUNT = 2
 SCAN_AUTOMERGE_MODE = True
 # </template>
 
-SCAN_POSTPROCESS_ENABLED_JOBS = [
-    'ralph.scan.postprocess.position',
-    'ralph.scan.postprocess.cache_price',
-]
+SCAN_POSTPROCESS_ENABLED_JOBS = []
 
 #
 # programmatic stuff that need to be at the end of the file
@@ -731,3 +731,6 @@ SCAN_PLUGINS = {
     },
 }
 RQ_TIMEOUT = 3000
+
+# url to page where user requests permission to module (eg. assets)
+# REQUEST_PERM_URL = 'http://tickets.office/request/ralph_module/permission'